Make all components use `position` and gridsize
This standardizes components like `pixi-volcano` to use `position` like
`pixi-city-container` does. This change also applies to sub-components
such as `GridLabel` and `GridSquare`.

Unlike `x` and `y` which are abiguous, `position` always refers to grid
coordinates.

The components also know the `gridSize` in order to draw themselves
in the correct location.

This will simplify the inversion of the grid in the next step.

[#165355691]