Update map and x-section to use @observer
Since these components are passed an observable array, they won't
re-render automatically when the array's internals update, and instead
only re-render when a primitive value updates. This means that their
rendering is always a step behind in the simulation.

This makes the less pure-componenty, but seems to be the recommended way.
See:

https://github.com/mobxjs/mobx-state-tree/issues/1060
https://github.com/mobxjs/mobx-react#faq
https://github.com/mobxjs/mobx/issues/101

An alternative would be to use `getSnapshot(data)`, but this is claimed
to be less efficient. (Untested, and perhaps for our use-case is negligible
and the cleaner code would be preferred.)