Add grid painting functionality
This adds a new data array to the simulation store, `gridColors`, which
gets computed when `paintGrid` is called. This is separate from the
`data` array as we want a painted grid to exist independently of the
data (i.e. if a student calculates data, paints the grid, and re-
calculates new data, the grid should reflect the old data).

[#165289061]