Restrict VEI range in code
There is no way to enforce a min or max on a block with a variable input.

The best we can do is simply clip the values behind the scenes, and tell
the user in the field that there is a range.

cf. blockly's own color function, which allows `Number` inputs for r,g,b,
and then clips them in the code:

https://github.com/google/blockly/blob/d12374b3fbbc2b45579d738c2360856aa74000e6/generators/javascript/colour.js#L50