Add ability for simulation code to pause blockly run
Simulation code can call `pause` and `unpause` to temporarily halt the
stepping of the next blocky block. This will be useful for allowing
certain blocks to take longer, e.g. if they want to display an animation.

This feature is not used yet, but can be tested by uncommenting the
two lines in `erupt`.

This could also be exposed to the interpreter, so that blocks themselves
could call pause and unpause, but it seems generally better for the logic
to be in the simulation.

[#165912766][#165312034]