Add animation flag and timing to simulation model
This doesn't include any animation code, but adds the property `isErupting`
to the simulation, and the various code to turn it off at the right time
(either on a timer when running or when we step to the next block or end
the run).

When `isErupting` is true, and we have an animation layer, we will run the
animation.

This also adds an Animation checkbox to the erupt block.

[#165312034]