Add animation layer, default configs, volcano image
The new component, `pixi-volcano-emitter`, uses `pixi-particles` as the
animation engine, and turns on and off a pre-defined particle animation
based on the simulation's `isErupting` property.

This includes a lava animation configuration, which we are not currently
using but is left in in case we change our minds, and to highlight that
new animations can be added to the existing map.

This also adds a quick drawing of a volcano seen from the top.

[#165312034]