Add plotting block, plot calculated values
This adds a new block to plot values on the graph, with explicit X and
Y set by the user from a list of options.

Each time the block is run, another point will be added.

The chart gets its x and y axis labels from the last point added. If a
point changes what is being measured, the chart is cleared.

This also adds a `vei` property for the simulation, which is used only
by this chart (for the user that sets a mass, but wants to plot against
VEI).

[#165289831]