Move all authoring options to models, rm authoring
This removes the local state of "authoring options" from app.tsx, and
places all of these options directly in the UI or Simulation models.

In the future, authoring the model will involve updating the models
directly.

However, for now this just rips out authoring, to be re-added in a
future commit, in order to attempt to simplify a messy refactoring.

Also update some of the simulation model defaults with the old authoring
options defaults.