Returns authoring menu and loads props into store
This changes the earlier authoring menu such that it is hydrated
directly from the current store state, and updates the store directly
whenever its values are changed.

(Previously, the authoring menu was tied to the component state of
app.tsx, and the store got updated as the various properties were passed
down to subcomponents.)

This also removes the initial slider values, as they will be set directly
from the model state in the next commit.

The stores props get updated wholesale. Currently all of these props can
be set without needing other side-effects by invoking individual actions,
but if that needs to change in the future we'll have to have a more
complex update method.

This introduces some typescript finessing to pick out which props we're
expecting and updating. This pattern will get expanded as we introduce
other sets of the props pulled out so far to support student and author
state saving to LARA.