Add new tephra3 thickness calculations; + tests
This implements the tephra model from Robert Constantinescu's python
simulation. This model integrates over multiple phi classes, as well
as a large eruption disk.

Tests are added which use output values from the python program to
confirm accuracy.

The tests are performed using the original high accuracy settings
(15 phi classes and an eruption cell size of 1000), but, for the sake
of speed, the default values used in the model are lower accuracy
(7 phi classes and a cell size of 3000). These values are fairly close,
you can see a comparison at
https://docs.google.com/spreadsheets/d/12EEJpJAnqk5pMMoa4cdtDSqUv7uiRPIPAGEIev1SJfQ/edit?usp=sharing

As we no longer are passing in particle size, the various components
that call this are updated.

The old model (tephra2) is maintained in this commit to make it easier
to go back and compare outputs, in the next commit it will be removed
entirely.