Make VEI an entirely computed value
This removes `vet` and `stagingVei` from the simulations-store's
properties, and instead makes it a computed value given mass and column
height.

If the user sets VEI, they will simply set the mass and column height.

If the user requests the VEI, we will make our best guess as to the VEI
given the mass and column height. If those properties were set using
`setVEI`, our "guess" will always be correct. If the user manually sets
mass and height to some non-possible combination, we will return a value
in-between the values suggested by the two properties.

This also adds `stagingVei` based on the staging mass and height values,
so that the vei slider updates as the user updates the mass and height
sliders.