Save data only after we change data for the first time
This should both fix the authoring issue where LARA reports that an
author has changed the authoring data (and so throwing up the "Leave
site?" dialog every time an author leaves the page), and will also stop
the student from saving ghost data, which may make loading authoring
data and linked data more confusing.

We compare the initial state with what is being saved back. After we
save once we'll always save, so as to avoid confusion when an author
resets to initial conditions.