Fix block inputs disappearing in fullscreen mode
The issue was that we expand the app's DOM element to fullscreen, but
blocky appends its own input dialogs outside of the wrapper element, at
the end of the document body. So those elements are not part of the
fullscreen app.

Since the model takes up the entire body, there should be little
difference expanding the app or the body. When we embed the model in an
iframe, only the app's document is expanded.

The only issue would be if we ever directly inlined the app in something
else, but this app is not designed for that.