Add D3-based scatter and radial charts
This adds two nearly-identical versions of a scatter chart, and a radial
chart for use with the wind data.

The scatter charts are implemented in both SVG and Canvas. The canvas
one is significantly faster for showing large quantities of data (>10k),
but the svg one may be better if we are going to include mouse
interactions, and so is kept for the time being.