Handle Dates in ChartStore; add more demo examples
This puts a little more of the data logic in the chart store itself, and
allows data sets to be made up of a mixture of numbers and Dates.

Two scatter charts are now available in the demos, one across all dates,
and one sorted by month.

Set the chart panel to automatically switch between canvas and svg scatter
plots based on number of points. Below ~5000 both charts render in about
the same time, at least on my machine.

Embiggen the canvas area slightly to allow us to see points directly on
axes (this is a problem in the month chart).