Extract blocky execution from simulation into controller
This extracts all code related to running or stepping through the blocky
code, and the creation of the interpreter, from the simulation store and
into its own controller. Among other things, this will make it cleaner
to pass in other stores.

I left the storing of the xmlCode in the simulation controller, which is
used for loading in state, as I didn't think it worth creating an MST
model specifically for the xml code and need to introduce a data
migration for all the existing authored models and student data. That
said, perhaps it would make sense to extract that as well and have the
migration after all.