Add blocks and functions for graphing all wind data
This adds the simplest blocks for graphing the wind data: a block that
returns all wind data with the Blockscript typing "Dataset", and a block
that will create a default time-wind speed scatter plot for any dataset
passed in.

This also adds a helper function for creating a default date-vs-y scatter
plot in the charts controller, so we don't need to add this code to the
interpreter.

Finally, this also resets the charts when the user hits the Play or Reset
buttons.

[#171186629]