Fix interpreter unwrapping code so it recurses
The functions created by `createNativeFunction` have an odd structure
where objects are wrapped by other objects, which put their properties
in a `data` or `properties` key. It turns out that this wrapping is
applied to all subjects as well, which meant that our current code
didn't allow passing nested objects from blocky to the interpreter
functions.

For the filter block, it will be nice to pass an object representing the
filter, so this makes that possible.