Add create- and add-to- samples collections blocks
This allows the user to define a "samples collection" at  given location
with a given name.

It then lets the user use that name in the "add to samples collection"
block. This is done by parsing the generated code each time the code
changes.

(Another way to do this would be to use Blockly's custom variables system,
with a dynamic fly-out that dynamically generates blocks based on the
variables created, but this was a massive headache.)

Note that if a user pulls out an "add to.." block, and then changes the
name of a collection (or pulls out an "add to..." block before a
collection is created), the previous value will remain selected until
changed. However, when the user goes to change it only valid collection
names will appear in the dropdown.