Throw error and stop program if missing block value
If the user runs a block that sets a tephra simulation parameter with
a missing value, or tries to use a block that requires a wind data set
with a missing value, or references a non-existant collection, we will
stop the execution of the code and alert the user with a native alert
box.

While the error does not specify the block with the problem, the block
will remain highlighted while the error is showing.

[#170927060]

 - throw error

 - throw