Create explicit locations before sample collections
This splits up the process of creating a samples collection into two
steps: creating a location, and creating a collection associated with
that location.

Both the collection and location have their own names. Trying to create
a collection without a valid location will throw a user-facing error.

The remaining code "just works" because we can use views to get a
collection's x and y location.

[#171749144]