Fix updating authoring state when menu changes
Previously, every time the authoring menu updated, we would load the
application store with the properties from the authoring panel, and
as a result we would throw all other state away.

This had the result (among other issues) that if an author updated a
slider in the control panel and then hid the slider, the sliders would
all reset to the initial values.

This fixes that problem by merging the current state of the application
into the authoring panel props and loading the application from that
merge.

[#169983751]