Add basic filtering using functions from wind data
This re-uses the DataSet functions, which work as-is with the GPS Station
data.

At some point we should think about integrating the GPS Station data with
the DataSet code more directly, allowing it to be one of the named data
sets, but it will be clearer how to do this once we have another GPS
dataset (position) to allow us to genericize properly.