Cache previous values for horizontal displacement
To calculate the displacement at year 1000, we needed to sum all the
displacements from year 0 to year 1000. To then calculate the displacement
at year 1500, we would then do the same thing. This was obviously
inefficient.

This introduces a cache for every original X location, with
year/displacement tuples for every previously-calculated value. When
calculating a value for any given year, we can look up the closest year
calculated before this year, for a given origin and speed, and step from
there.

A typical cache size will be about ~300 separate tuples calculated, times
~30 vertical lines and 3 stations, caching a total of ~20,000 numbers,
which isn't a significant overhead.