Move rotation of filtered wind into interpreter
This is functionally identical to the previous version, and still only
rotates numerical data (e.g. "direction: 20").

This has two advantages: first, if the Code tab is enabled, the code
will look like what the user expects to see (e.g. if you put "20" in
the direction slot, you'll get a filter with {direction: 20}), and
second it will be much easier to deal with the range filter, which will
be in the form of the string "(({min: x, max: y}))" before it's passed
to the interpreter.