Prevent modifying data point when we get single sample
Previously if we used this block, the original data point would get
modified. This was most obvious if we filtered to get a single data point
(e.g. direction: 4), and then tried to get the same data point again
later, as it would no longer be found by the filter.