Fix x-y coords reversal in LatLngToLocal, mirroring
The tephra thickness layer map has been mirrored along the x=y diagonal
axis since it was first created. This was hidden by then also mirroring
the underlying tephra calculation. The result of this double-mirroring
was that the tephra-thickness-layer-map looked correct to the eye.
However, when we actually went to ask for the tephra thickness at known
locations, we were getting incorrect values, because the underlying data
was being mirrored.

After much head-bashing, I finally figured out that the source of the
tephra-thickness-layer-map mirroring was this code here. This is only
used by the map layer, and the obsolete cross-section view, and is used
to calculate the displacement from each point in the map to the volcano.

This is an easy mistake to make because we say both "x, y" and "lat, long",
though they are inverse. (The fact that the underlying calc then had to
be mirrored could have been a clue, though...)

Tephra calc will be fixed in next commit. This brings the view in line
with what we are actually calculating.