Fix the tephra calculation so it does not mirror
After the tephra-thickness-layer was accidentally mirrored along the
x=y axis, the tephra calculation was mirrored to compensate by
subtracting the rotation from 360º.

This returns the tephra calc to its original version, which assumes
an easterly wind (i.e. tephra travels along the x axis), and then
rotates the grid to simulate a specific wind direction.

(The old comment "...which assumes a wind due South" was written after
this mirroring was added, and was not correct. Likewise, the test was
written after this.)