Add migration to convert author/student state to latest
This creates a framework for a series of migrations that can convert any
old state to the latest version. At the moment, this only converts
from the pre-seismic-unit states (which is what's saved, as of this commit,
on production) to the post-seismic-unit state.

This belatedly bumps the serialization version number to 2. Because
post-seismic-unit work may have been saved under version number 1, the
migration duck-types to check if it needs migration.