Samples collection now take threshold but not location
There is no longer any explicit association between a collection and a
location. If the team ever wants risk triangles back, the "Show risk"
block will be modified to take a location.

Collections now keep their own threshold, which will be used any time
the risk is shown or a chart is made of the collection.

Two collections cannot be created with the same name, preventing users
from adding the block to a loop.