Save state when user runs; separate student/author states
This adds a `hasRunOnce` flag to the blocky controller and serialized
state, so that the user saves some data when the interact with a
pre-authored model, and this will show up in the report.

Because we don't want to save when the author has run the model, this
separates out author state and student state for the first time.

However, we also have to add this to the migration, so that when the
user loads the state for the first time, their addition of
`hasRunOnce: false` doesn't mark the store as dirty.