Fix issue caused by migrating null data
When an author first creates a model, it passes {} to the migration,
which doesn't deal with that correctly. This was causing `updateStores`
to throw an exception, preventing the store listeners from ever being
added.

This fixes the issue by making the initial type of the data loaded from
LARA explicit such that it may be {}, and passes back a valid but empty
state if so.