Split horizontal lines in two and calculate more accurately
Previously we drew each horizontal line as a single curve. However, now
with the possibility of earthquakes splitting the line in two we need
to pass back two separate lines for each horizontal line.

Because of the way the Bezier curve works, this would make the curved
sections at the center more jagged than before. So we now sample more
points near the fault line.