Add earthquake model
This simply defines "friction" as the max displacement a plate can move
before an earthquake will shift the entire plate so that the distorted
area under strain is flat with respect to the rest of the plate. This
looks visually like the curved lines snapping to become straight with
the horizontal portion of the lines.