Add new version of logic_compare that accepts code
The built-in logic_compare block, which provides ===, <, etc comparisons,
doesn't work well when handling code blocks that return numbers, because
of the way the js interpreter wraps the value of the results.

By creating our own version, it's now possible to mix-and-match code blocks
and pure numbers and compare them as expected.

(This was a very frustrating head-banger.)