Add a method to force Blockly to reload
We had previously designed the blockly controller to trigger a rerender
of the component when the initial XML code changes. When we want to
reload to the initial conditions, the initialXmlCode does not change.

This adds a dummy property that the component can watch, which when
changed will trigger Blockly to reload from scratch.

(This is a little hacky, but is much less hacky than my previous four
solutions...)