181021708: Rename *Strain-* components to *Deformation-*
We have changed the text labels of the block to use `Deformation build-up` instead of `Strain rate`, but we did not rename the variables or components behind the scenes.

This story will make it easier for developers to find the correct blocks in the source code.

These changes are being done in their own story because it is difficult to track file renaming in diffs. It will make code review easier to do them after the other changes are accepted.

[#181021708]
https://www.pivotaltracker.com/story/show/181021708